add pv, move tree node into go_comm
